+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "data-modela",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "description": "an application that uses sequelize methods to work with dummy-data when building appliication",
   "main": "./lib/main/index.js",
   "scripts": {
@@ -11,7 +11,7 @@
     "prepublish": "npm test",
     "start": "nodemon --exec npm run babel:node",
     "js:lint": "eslint src/",
-    "coverage2": "nyc npm run test && nyc report --reporter=text-lcov | coveralls",
+    "coverage2": "nyc npm run test && nyc report --reporter=lcov",
     "coverage": "istanbul cover ./node_modules/mocha/bin/_mocha -- -R spec ./lib/test/",
     "send:coverage": "istanbul cover ./node_modules/mocha/bin/_mocha --report lcovonly -- -R spec ./lib/test/ && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
   },
